requirements.txt

python-telegram-bot==20.3
pyrogram==2.0.106
motor==3.1.1
asyncio
